+++ app/controllers/api/members_controller.rb
@@ -180,7 +180,7 @@ class API::MembersController < API::ApiController
   end
 
   def search
-    @members = Members::ListService.search(current_user, params[:query], params[:subscription], params[:project])
+    @members = Members::ListService.search(current_user, params[:query], params[:subscription], params[:include_admins])
   end
 
   def mapping