+++ test/integration/credits/user_info_test.rb
@@ -1,0 +1,15 @@
+require 'minitest/autorun'
+
+class UserInfoTest < Minitest::Test
+  def setup
+    # Do nothing
+  end
+
+  def teardown
+    # Do nothing
+  end
+
+  def test
+    skip 'Not implemented'
+  end
+end