+++ app/frontend/src/javascript/models/credit.ts
@@ -1,0 +1,22 @@
+import { TDateISO } from '../typings/date-iso';
+
+export type CreditableType = 'Training' | 'Machine' | 'Space';
+
+export interface Credit {
+  id?: number,
+  creditable_id: number,
+  creditable_type: CreditableType,
+  created_at?: TDateISO,
+  updated_at?: TDateISO,
+  plan_id?: number,
+  hours: number,
+  creditable?: {
+    id: number,
+    name: string
+  },
+  used_credits?: [{
+    id: number,
+    created_at: TDateISO,
+    hours_used: number
+  }]
+}