+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next deploy
 
+- Fix a bug: unable to reserve a space
 ## v5.4.7 2022 June 29
 
 - Removed the admins' button to scroll to the featured event