+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next deploy
 
+- My reservations dashboard
 - Fix a bug: unable to reserve a space
 - Added a test case for space reservation
 ## v5.4.7 2022 June 29